refactor: move analytics component up a level
