+++ apps/journeys-admin/src/components/Editor/EditToolbar/Analytics/Analytics.spec.tsx
@@ -4,8 +4,8 @@ import { SnackbarProvider } from 'notistack'
 
 import { JourneyProvider } from '@core/journeys/ui/JourneyProvider'
 
-import { TeamProvider } from '../../../../Team/TeamProvider'
-import { defaultJourney } from '../../../data'
+import { TeamProvider } from '../../../Team/TeamProvider'
+import { defaultJourney } from '../../data'
 
 import { Analytics } from './Analytics'
