+++ apps/journeys-admin/src/components/Editor/EditToolbar/Menu/Menu.tsx
@@ -21,9 +21,9 @@ import { Role } from '../../../../../__generated__/globalTypes'
 import { setBeaconPageViewed } from '../../../../libs/setBeaconPageViewed'
 import { DuplicateBlock } from '../../../DuplicateBlock'
 import { MenuItem } from '../../../MenuItem'
+import { Analytics } from '../Analytics'
 import { DeleteBlock } from '../DeleteBlock'
 
-import { Analytics } from './Analytics'
 import { CopyMenuItem } from './CopyMenuItem'
 import { CreateTemplateMenuItem } from './CreateTemplateMenuItem'
 import { LanguageMenuItem } from './LanguageMenuItem'
